bddframework/pkg/framework/kogitoservice/app_configmap_test.go
// Copyright 2020 Red Hat, Inc. and/or its affiliates
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

package kogitoservice

import (
	"github.com/kiegroup/kogito-operator/core/operator"
	"github.com/kiegroup/kogito-operator/core/test"
	"github.com/kiegroup/kogito-operator/meta"
	"github.com/stretchr/testify/assert"
	"testing"
)

func TestCreateAppConfigMap(t *testing.T) {
	service := test.CreateFakeDataIndex(t.Name())
	cli := test.NewFakeClientBuilder().Build()
	context := operator.Context{
		Client: cli,
		Log:    test.TestLogger,
		Scheme: meta.GetRegisteredSchema(),
	}
	appConfigMapHandler := NewAppConfigMapHandler(context)
	appProps := map[string]string{
		"key1": "value1",
	}
	configMap := appConfigMapHandler.CreateAppConfigMap(service, appProps)
	assert.NotNil(t, context)
	assert.Equal(t, 1, len(configMap.Data))
}
